+++ test/authenticator.middleware.test.js
@@ -150,6 +150,51 @@ describe('Authenticator', function() {
       });
     });
     
+    describe('handling a request with instantiated strategy', function() {
+      function Strategy() {
+      }
+      Strategy.prototype.authenticate = function(req) {
+        var user = { id: '1', username: 'jaredhanson' };
+        this.success(user);
+      };
+
+      var passport = new Authenticator();
+
+      var request, error;
+
+      before(function(done) {
+        chai.connect.use(passport.authenticate(new Strategy()))
+          .req(function(req) {
+            request = req;
+
+            req.logIn = function(user, options, done) {
+              this.user = user;
+              done();
+            };
+          })
+          .next(function(err) {
+            error = err;
+            done();
+          })
+          .dispatch();
+      });
+
+      it('should not error', function() {
+        expect(error).to.be.undefined;
+      });
+
+      it('should set user', function() {
+        expect(request.user).to.be.an('object');
+        expect(request.user.id).to.equal('1');
+        expect(request.user.username).to.equal('jaredhanson');
+      });
+      
+      it('should set authInfo', function() {
+        expect(request.authInfo).to.be.an('object');
+        expect(Object.keys(request.authInfo)).to.have.length(0);
+      });
+    });
+    
   });
